InfinityConstellation-CSA-Agent/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
streamlit==1.37.0

# AI & Agents
openai>=1.40.0
anthropic>=0.34.0
mistralai>=0.3.0
google-generativeai>=0.7.0
langchain>=0.2.7
langgraph>=0.0.32
deepagents>=0.1.5
massgen==0.0.5

# Tools
tavily-python>=0.1.0
agentql>=0.1.0
jigsawstack>=0.2.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.29
alembic>=1.13.1

# RAG & Vector DB
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0

# Testing
pytest>=8.2.0
httpx>=0.27.0